+++ esp-rtos/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Interrupt-safe semaphore operations (#4396)
 
 ### Changed